refactor: CLI cleanup — ValueEnum types, zero clippy warnings, dead code audit
Replace string CLI args with clap ValueEnum enums (IngestFormat, CsvFormat,
OutputFormat, DocFormat, PreprocessFormat, EnergyLevel, GtdState, ParaCategory,
ProactivityLevel) for compile-time validation and tab-completion. Add
value_delimiter=',' to 14 comma-separated args eliminating manual split().

Fix server detection inconsistency: AgentAction::Chat and AgentAction::Repl
now try discover_server() before falling back to local engine, matching
Commands::Chat behavior.

Eliminate all 330 clippy warnings (137 collapsible ifs using edition 2024
let-chains, 51 redundant closures, 15 contains→iter().any(), and ~130 others
across 80+ files). Replace 60+ error-swallowing `let _ =` patterns with
tracing::warn!/debug! logging. Add justification comments to all
#[allow(dead_code)] annotations. Fix pre-existing SymbolKind import in
microtheory.rs tests.

All 1,677 tests pass, zero clippy warnings from project code.

+++ src/vsa/grounding.rs
@@ -145,7 +145,7 @@ fn blend_vectors(
     let mut result = base_data.to_vec();
 
     // Deterministic PRNG seeded from symbol ID (xorshift64)
-    let mut state = u64::from(symbol.get()).wrapping_mul(0x517cc1b727220a95) | 1;
+    let mut state = symbol.get().wrapping_mul(0x517cc1b727220a95) | 1;
     let threshold = (weight * 256.0) as u64;
 
     for (i, byte) in result.iter_mut().enumerate() {
